// src/app/pages/Task Management/Models/QuestionMaster.ts
export class QuestionMaster {
  ID: any;
  QUESTION_TYPE: any = '';
  CHAPTER_ID: number;
  QUESTION: any;
  QUESTION_IMAGE: any;
  DESCRIPTION: any;
  ANSWER: any;
  ANSWER_IMAGE: any;
  MARKS: any;
  SEQ_NO: number = 0;
  STATUS: boolean = true;
  OPTIONS = [];
  CLASS_ID: number = 0;
  SCHOOL_ID: number = 0;
  SUBJECT_ID: number = 0;

  DIVISION_ID: number;
  CLIENT_ID: any;
}